Removed seperate logout tests
